Revert "Extract chart to own component file"
This reverts commit 5757fc9f4856a28b13b37a28baa38690db4ebe3a.
